+++ src/main/java/pairmatching/domain/pair/Pair.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 import pairmatching.domain.course.Course;
 import pairmatching.domain.crew.Crew;
-import pairmatching.domain.crew.Crews;
+import pairmatching.repository.Crews;
 import pairmatching.domain.level.Level;
 import pairmatching.domain.mission.Mission;
