+++ src/main/java/pairmatching/repository/Crews.java
@@ -1,9 +1,10 @@
-package pairmatching.domain.crew;
+package pairmatching.repository;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 import pairmatching.domain.course.Course;
+import pairmatching.domain.crew.Crew;
 
 public class Crews {
